src/components/profile: fix invite team members btn element layout position
