+++ src/components/profile/TeamMembersList.vue
@@ -12,17 +12,22 @@
  */
 
 // libraries
-import { defineComponent, computed } from 'vue';
+import { defineComponent, computed, ref } from 'vue';
 
 // composables
 import { useTeamMemberApprovalStatus } from '../../composables/useTeamMemberApprovalStatus';
 import { i18n } from '../../boot/i18n';
+
+// Components
+import DialogDefault from '../global/DialogDefault.vue';
+import FormInviteToTeam from './FormInviteToTeam.vue';
 
 // config
 import { rideToWorkByBikeConfig } from '../../boot/global_vars';
 
 // stores
 import { useRegisterChallengeStore } from '../../stores/registerChallenge';
+import { useChallengeStore } from '../../stores/challenge';
 
 // types
 import type { ExtendedMemberResults } from '../types/Results';
@@ -32,8 +37,13 @@ import { getGenderLabel } from '../../utils/get_gender_label';
 
 export default defineComponent({
   name: 'TeamMembersList',
+  components: {
+    DialogDefault,
+    FormInviteToTeam,
+  },
   setup() {
     const registerChallengeStore = useRegisterChallengeStore();
+    const challengeStore = useChallengeStore();
 
     const teamMembers = computed<ExtendedMemberResults[]>(
       (): ExtendedMemberResults[] => {
@@ -48,6 +58,18 @@ export default defineComponent({
 
     const { borderRadiusCard: borderRadius } = rideToWorkByBikeConfig;
 
+    const isDialogInviteTeamMembersOpen = ref(false);
+
+    const remainingSlots = computed<number>((): number => {
+      const myTeam = registerChallengeStore.getMyTeam;
+      if (!myTeam) return 0;
+
+      const maxTeamMembers = challengeStore.getMaxTeamMembers;
+      if (!maxTeamMembers) return 0;
+
+      return maxTeamMembers - myTeam.member_count;
+    });
+
     return {
       borderRadius,
       getGenderLabel,
@@ -55,6 +77,8 @@ export default defineComponent({
       getStatusColor,
       getStatusIcon,
       i18n,
+      isDialogInviteTeamMembersOpen,
+      remainingSlots,
       teamMembers,
     };
   },
@@ -70,21 +94,69 @@ export default defineComponent({
     >
       {{ $t('teamMembersList.titleMembers') }}
     </h3>
+
     <!-- List -->
     <q-expansion-item header-class="disable-bg-hover" data-cy="team-members">
       <template v-slot:header>
-        <q-item-section avatar>
-          <q-avatar icon="groups" color="secondary" text-color="primary" />
-        </q-item-section>
+        <div class="row full-width items-center">
+          <div class="col-9">
+            <!-- Show hide expansion team members element -->
+            <div class="row items-center">
+              <div class="col-auto q-mr-lg">
+                <!-- Icon -->
+                <q-avatar
+                  icon="groups"
+                  color="secondary"
+                  text-color="primary"
+                />
+              </div>
+              <!-- Title -->
+              <div class="col-auto">
+                {{ $t('teamMembersList.showHideMembers') }}
+              </div>
+            </div>
+          </div>
 
-        <q-item-section>
-          {{ $t('teamMembersList.showHideMembers') }}
-        </q-item-section>
+          <div class="col-xl-3 col-lg-3 col-auto">
+            <q-item-section avatar>
+              <!-- Invite new team members btn -->
+              <q-chip
+                icon="group_add"
+                color="secondary"
+                text-color="primary"
+                size="md"
+                :class="{ 'q-mt-md': $q.screen.lt.md }"
+                :disable="remainingSlots === 0"
+                @click.stop="isDialogInviteTeamMembersOpen = true"
+                clickable
+                data-cy="profile-details-invite-team-members"
+                >{{ $t('profile.inviteTeamMembers') }}</q-chip
+              >
+              <!-- Invite new team members modal dialog  -->
+              <dialog-default
+                v-model="isDialogInviteTeamMembersOpen"
+                data-cy="profile-details-invite-team-members-dialog"
+              >
+                <!-- Dialog title -->
+                <template #title>
+                  {{ $t('profile.inviteTeamMembers') }}
+                </template>
+                <!-- Dialog content -->
+                <template #content>
+                  <form-invite-to-team
+                    :remaining-slots="remainingSlots"
+                    :on-close="() => (isDialogInviteTeamMembersOpen = false)"
+                  />
+                </template>
+              </dialog-default>
+            </q-item-section>
+          </div>
+        </div>
       </template>
       <q-card>
         <q-card-section>
           <q-list bordered separator class="rounded-borders">
-            <!-- Member -->
+            <!-- Team member expansion item element -->
             <q-expansion-item
               class="full-width"
               v-for="member in teamMembers"
@@ -96,6 +168,7 @@ export default defineComponent({
                 <div class="row full-width">
                   <div class="col-9">
                     <q-item-section avatar>
+                      <!-- Icon -->
                       <q-chip
                         color="secondary"
                         text-color="primary"
@@ -132,6 +205,7 @@ export default defineComponent({
                 <q-card-section>
                   <!-- Email -->
                   <q-chip color="secondary" text-color="primary">
+                    <!-- Icon -->
                     <q-avatar
                       icon="contact_mail"
                       color="secondary"
