+++ src/i18n/cs.toml
@@ -696,7 +696,7 @@ buttonDeleteAccount = "Smazat"
 buttonDownloadInvoice = "Stáhnout potvrzení o platbě"
 buttonFillQuestionnaire = "Vyplnit dotazník"
 descriptionNickname = "Zobrazí se ve veřejných výsledcích místo vašeho jména"
-inviteTeamMembers = "Pozvat členy do týmu"
+inviteTeamMembers = "Pozvat členy"
 labelAddressSubsidiary = "Adresa / pobočka:"
 labelDeleteAccount = "Smazat účet a všechny příslušné údaje"
 labelDeleteAccountDescription = "Smazáním účtu smažete všechny vaše údaje. Přejete si pokračovat?"
