+++ src/i18n/en.toml
@@ -694,7 +694,7 @@ buttonDownloadInvoice = "Download payment confirmation"
 buttonFillQuestionnaire = "Fill in questionnaire"
 buttonRemoveEmailField = "Remove email"
 descriptionNickname = "Appears in public results instead of your name"
-inviteTeamMembers = "Invite team members"
+inviteTeamMembers = "Invite members"
 labelAddressSubsidiary = "Address / Subsidiary:"
 labelDeleteAccount = "Delete account and all associated data"
 labelDeleteAccountDescription = "Deleting your account will delete all your data. Do you want to continue?"
