+++ src/i18n/sk.toml
@@ -694,7 +694,7 @@ buttonDeleteAccount = "Zmazať"
 buttonDownloadInvoice = "Stiahnutie potvrzenia o platbe"
 buttonFillQuestionnaire = "Vyplniť dotazník"
 descriptionNickname = "Zobrazuje sa vo verejných výsledkoch namiesto vášho mena"
-inviteTeamMembers = "Pozvať členov do tímu"
+inviteTeamMembers = "Pozvať členov"
 labelAddressSubsidiary = "Adresa / pobočka:"
 labelDeleteAccount = "Zmazať účet a všetky príslušné údaje"
 labelDeleteAccountDescription = "Zmazaním účtu zmažete všetky vaše údaje. Prajete si pokračovať?"
